Rename PacketSendFilterManager files
